refactor(#29): updated readme, rsa fixes.
